refactor log printing and shell
issues: https://github.com/zzxyb/wsm/issues/1
