+++ shell/wsm_layer_shell.h
@@ -65,16 +65,6 @@ struct wsm_layer_surface {
     struct wlr_layer_surface_v1 *layer_surface;
 };
 
-struct wsm_layer_popup {
-    struct wlr_xdg_popup *wlr_popup;
-    struct wlr_scene_tree *scene;
-    struct wsm_layer_surface *toplevel;
-
-    struct wl_listener destroy;
-    struct wl_listener new_popup;
-    struct wl_listener commit;
-};
-
 struct wsm_layer_shell {
     struct wlr_layer_shell_v1 *wlr_layer_shell;
     struct wl_listener layer_shell_surface;
