+++ shell/wsm_xdg_shell.c
@@ -23,6 +23,7 @@ THE SOFTWARE.
 */
 
 #include "wsm_xdg_shell.h"
+#include "wsm_xdg_popup.h"
 #include "wsm_server.h"
 #include "wsm_log.h"
 #include "wsm_view.h"
@@ -55,118 +56,6 @@ THE SOFTWARE.
 
 #define WSM_XDG_SHELL_VERSION 5
 #define CONFIGURE_TIMEOUT_MS 100
-
-static struct wsm_xdg_popup *popup_create(
-    struct wlr_xdg_popup *wlr_popup, struct wsm_view *view,
-    struct wlr_scene_tree *parent);
-
-static void popup_handle_new_popup(struct wl_listener *listener, void *data) {
-    struct wsm_xdg_popup *popup =
-        wl_container_of(listener, popup, new_popup);
-    struct wlr_xdg_popup *wlr_popup = data;
-    popup_create(wlr_popup, popup->view, popup->xdg_surface_tree);
-}
-
-static void popup_handle_destroy(struct wl_listener *listener, void *data) {
-    struct wsm_xdg_popup *popup = wl_container_of(listener, popup, destroy);
-
-    wl_list_remove(&popup->new_popup.link);
-    wl_list_remove(&popup->destroy.link);
-    wl_list_remove(&popup->surface_commit.link);
-    wl_list_remove(&popup->reposition.link);
-    wlr_scene_node_destroy(&popup->scene_tree->node);
-    free(popup);
-}
-
-static void popup_unconstrain(struct wsm_xdg_popup *popup) {
-    struct wsm_view *view = popup->view;
-    struct wlr_xdg_popup *wlr_popup = popup->wlr_xdg_popup;
-
-    struct wsm_workspace *workspace = view->container->pending.workspace;
-    if (!workspace) {
-        // is null if in the scratchpad
-        return;
-    }
-
-    struct wsm_output *output = workspace->output;
-
-    // the output box expressed in the coordinate system of the toplevel parent
-    // of the popup
-    struct wlr_box output_toplevel_sx_box = {
-        .x = output->lx - view->container->pending.content_x + view->geometry.x,
-        .y = output->ly - view->container->pending.content_y + view->geometry.y,
-        .width = output->width,
-        .height = output->height,
-    };
-
-    wlr_xdg_popup_unconstrain_from_box(wlr_popup, &output_toplevel_sx_box);
-}
-
-static void popup_handle_surface_commit(struct wl_listener *listener, void *data) {
-    struct wsm_xdg_popup *popup = wl_container_of(listener, popup, surface_commit);
-    if (popup->wlr_xdg_popup->base->initial_commit) {
-        popup_unconstrain(popup);
-    }
-}
-
-static void popup_handle_reposition(struct wl_listener *listener, void *data) {
-    struct wsm_xdg_popup *popup = wl_container_of(listener, popup, reposition);
-    popup_unconstrain(popup);
-}
-
-static struct wsm_xdg_popup *popup_create(struct wlr_xdg_popup *wlr_popup,
-                                           struct wsm_view *view, struct wlr_scene_tree *parent) {
-    struct wlr_xdg_surface *xdg_surface = wlr_popup->base;
-
-    struct wsm_xdg_popup *popup = calloc(1, sizeof(struct wsm_xdg_popup));
-    if (!popup) {
-        return NULL;
-    }
-
-    popup->wlr_xdg_popup = wlr_popup;
-    popup->view = view;
-
-    popup->scene_tree = wlr_scene_tree_create(parent);
-    if (!popup->scene_tree) {
-        free(popup);
-        return NULL;
-    }
-
-    popup->xdg_surface_tree = wlr_scene_xdg_surface_create(
-        popup->scene_tree, xdg_surface);
-    if (!popup->xdg_surface_tree) {
-        wlr_scene_node_destroy(&popup->scene_tree->node);
-        free(popup);
-        return NULL;
-    }
-
-    popup->desc.relative = &view->content_tree->node;
-    popup->desc.view = view;
-
-    if (!wsm_scene_descriptor_assign(&popup->scene_tree->node,
-                                 WSM_SCENE_DESC_POPUP, &popup->desc)) {
-        wsm_log(WSM_ERROR, "Failed to allocate a popup scene descriptor");
-        wlr_scene_node_destroy(&popup->scene_tree->node);
-        free(popup);
-        return NULL;
-    }
-
-    popup->wlr_xdg_popup = xdg_surface->popup;
-    struct wsm_xdg_shell_view *shell_view =
-        wl_container_of(view, shell_view, view);
-    xdg_surface->data = shell_view;
-
-    wl_signal_add(&xdg_surface->surface->events.commit, &popup->surface_commit);
-    popup->surface_commit.notify = popup_handle_surface_commit;
-    wl_signal_add(&xdg_surface->events.new_popup, &popup->new_popup);
-    popup->new_popup.notify = popup_handle_new_popup;
-    wl_signal_add(&wlr_popup->events.reposition, &popup->reposition);
-    popup->reposition.notify = popup_handle_reposition;
-    wl_signal_add(&wlr_popup->events.destroy, &popup->destroy);
-    popup->destroy.notify = popup_handle_destroy;
-
-    return popup;
-}
 
 static struct wsm_xdg_shell_view *xdg_shell_view_from_view(
     struct wsm_view *view) {
@@ -233,8 +122,6 @@ static void set_tiled(struct wsm_view *view, bool tiled) {
         }
         wlr_xdg_toplevel_set_tiled(view->wlr_xdg_toplevel, edges);
     } else {
-        // The version is too low for the tiled state; configure as maximized instead
-        // to stop the client from drawing decorations outside of the toplevel geometry.
         wlr_xdg_toplevel_set_maximized(view->wlr_xdg_toplevel, tiled);
     }
 }
@@ -325,7 +212,6 @@ static void handle_commit(struct wl_listener *listener, void *data) {
         if (view->xdg_decoration != NULL) {
             set_xdg_decoration_mode(view->xdg_decoration);
         }
-        // XXX: https://github.com/wsmwm/wsm/issues/2176
         wlr_xdg_surface_schedule_configure(xdg_surface);
         wlr_xdg_toplevel_set_wm_capabilities(view->wlr_xdg_toplevel,
                                              XDG_TOPLEVEL_WM_CAPABILITIES_FULLSCREEN);
@@ -345,13 +231,9 @@ static void handle_commit(struct wl_listener *listener, void *data) {
                     new_geo.y != view->geometry.y;
 
     if (new_size) {
-        // The client changed its surface size in this commit. For floating
-        // containers, we resize the container to match. For tiling containers,
-        // we only recenter the surface.
         memcpy(&view->geometry, &new_geo, sizeof(struct wlr_box));
         if (container_is_floating(view->container)) {
             view_update_size(view);
-            // Only set the toplevel size the current container actually has a size.
             if (view->container->current.width) {
                 wlr_xdg_toplevel_set_size(view->wlr_xdg_toplevel, view->geometry.width,
                                           view->geometry.height);
@@ -366,10 +248,6 @@ static void handle_commit(struct wl_listener *listener, void *data) {
         bool successful = transaction_notify_view_ready_by_serial(view,
                                                                   xdg_surface->current.configure_serial);
 
-        // If we saved the view and this commit isn't what we're looking for
-        // that means the user will never actually see the buffers submitted to
-        // us here. Just send frame done events to these surfaces so they can
-        // commit another time for us.
         if (view->saved_surface_tree && !successful) {
             view_send_frame_done(view);
         }
@@ -395,7 +273,7 @@ static void handle_new_popup(struct wl_listener *listener, void *data) {
         wl_container_of(listener, xdg_shell_view, new_popup);
     struct wlr_xdg_popup *wlr_popup = data;
 
-    struct wsm_xdg_popup *popup = popup_create(wlr_popup,
+    struct wsm_xdg_popup *popup = wsm_xdg_popup_create(wlr_popup,
                                                 &xdg_shell_view->view, global_server.wsm_scene->layers.popup);
     if (!popup) {
         return;
@@ -582,7 +460,7 @@ void handle_xdg_shell_toplevel(struct wl_listener *listener, void *data) {
 
     struct wsm_xdg_shell_view *xdg_shell_view =
         calloc(1, sizeof(struct wsm_xdg_shell_view));
-    if (!wsm_assert(xdg_shell_view, "Failed to allocate view")) {
+    if (!wsm_assert(xdg_shell_view, "Could not create wsm_xdg_shell_view: allocation failed!")) {
         return;
     }
